Move Aplicacao para uma nova pasta /app, e ajusta para usar RepositorioProdutoFactory
